creation of Professor MVC and complete implementation

+++ myapp/database/migrations/2025_09_27_083524_create_professors_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('first_name');
             $table->string('last_name');
             $table->enum('professor_type', ['regular', 'non-regular', 'none'])->default('regular');
-            $table->integer('max_unit_load');
+            $table->decimal('max_unit_load', 10, 1);
             $table->integer('professor_age')->nullable();
             $table->string('position')->nullable();
             $table->timestamps();
